Save email address after verifying the email address
In the current implementation when multi step authentication is enabled and if the user has not specify the email address in the profile, it will prompt user to enter email address and then directly saves it. If user enters an
incorrect email there is no way for the user to change it himself. This implementation will only save the email address once it is confirmed with the OTP code sends.